DistinctCountTranslationTests added. One test moved, one more added.
